+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.1.1] - 2026-03-23
+
+### Fixed
+- Fixed support of JWT authenticatiors without token-app-property. 
+- Fixed support for custom audience claim.
+
 ## [2.1.0] - 2026-01-12
 
 ### Changed
@@ -54,7 +60,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release
 
-[Unreleased]: https://github.com/cyberark/conjur-action/compare/v2.1.0...HEAD
+[Unreleased]: https://github.com/cyberark/conjur-action/compare/v2.1.1...HEAD
+[2.1.0]: https://github.com/cyberark/conjur-action/compare/v2.1.0...v2.1.1
 [2.1.0]: https://github.com/cyberark/conjur-action/compare/v2.0.12...v2.1.0
 [2.0.12]: https://github.com/cyberark/conjur-action/compare/v2.0.11...v2.0.12
 [2.0.11]: https://github.com/cyberark/conjur-action/compare/v2.0.10...v2.0.11